requirements.txt
requests>=2.31.0
pygame>=2.5.0
schedule>=1.2.0
pyautogui>=0.9.54
pillow>=10.3.0
python-docx>=1.1.0
openpyxl>=3.1.2
PyPDF2>=3.0.1
markdown>=3.5.2
opencv-python>=4.9.0.80
pywin32>=306
beautifulsoup4>=4.12.0
PyQt5>=5.15.0
pyperclip>=1.8.0
winotify>=1.1.0
keyboard>=0.13.5
vosk>=0.3.45
pyaudio>=0.2.13
numpy>=1.24.0
python-pptx>=0.6.21